Extract manifest sorting logic
