+++ pkg/types/manifest.go
@@ -1,0 +1,3 @@
+package types
+
+type Manifest = map[string]any
